sql: try to track non-pgerrors by source location
for non-pgerror errors, we can attempt to track them by where they came from before falling back to 'unknown'.

Release note: none.
